Moving scripting interfaces into the correct folders.
