+++ Assets/Source/Player/Scripting/Interface/ProximityChecker.cs
