+++ Assets/Source/Player/Scripting/Interface/TimerJsInterface.cs
