--- Assets/Source/Player/Scripting/Interop/MessageRouterScriptingInterface.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 74ad62de306a38a4eb69864f21a6c31e
-timeCreated: 1507240678
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
